+++ src/features/cheatcodes/pages/Navigation.tsx
@@ -10,8 +10,8 @@ import { useSignIn } from 'features/auth/api'
 import { DEEPLINK_DOMAIN } from 'features/deeplinks'
 import { openExternalUrl } from 'features/navigation/helpers'
 import { UseNavigationType } from 'features/navigation/RootNavigator'
-import { useDistance } from 'libs/geolocation/hooks/useDistance'
 import { AsyncError } from 'libs/errorMonitoring'
+import { useDistance } from 'libs/geolocation/hooks/useDistance'
 import { QueryKeys } from 'libs/queryKeys'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { ModalHeader } from 'ui/components/modals/ModalHeader'
