+++ src/features/offer/components/index.ts
@@ -2,6 +2,5 @@ export { AccordionItem } from './AccordionItem'
 export { CallToAction } from './CallToAction'
 export { OfferHeader } from './OfferHeader'
 export { OfferIconCaptions } from './OfferIconCaptions'
-export { WhereSection } from './WhereSection'
 export { AccessibilityBlock } from './AccessibilityBlock'
 export { OfferPartialDescription } from './OfferPartialDescription'
